lazy loading for infra assets (#62363)
